Updating the script to work with current git versions.
Main difference is that LLVM 3.4 also builds libc++ if found in the
projects directory so we don't need to build that manually anymore.
That however changes the build order a bit.

Also unifying the three build stages to reuse the same code with
different options.

Also includes some general cleanup and renaming of the options
selecting what to build.

The three build stage