Link libcxxabi statically into libcxx.
Simply by *not* installing the shared object version of libc++abi into the
install prefix, the subsequent stage1 build of libc++ picked up the static
version of libc++abi. This came quite surprisingly, because the libc++abi build
script did not even have support for a static version until we patched it. So
why would the libc++ build script then understand how to integrate a static
version of the library? My hunch is that CMake provided the missing glue.